// types.ts
export interface Branch {
  name: string;
  address: string;
  hours: string;
  phone: string;
  location: string;
}

export interface Service {
  id: string;
  title: string;
  description: string;
  icon: string;
}

export interface BatteryBrand {
  name: string;
  origin: string;
  flag: string;
}